FIX: rename/refactor eventbrite sources and tests (#175) :bug:
